Restricts parsing to Haskell files
It's possible for Haskell projects to have non-Haskell files in their
source trees for many good reasons. This restricts parsing to only files
that end in ".hs". Notably, this currently excluseds ".lhs" because we
see them so rarely and I don't have a quick test example to use at
the moment.