Only shows imports of local modules in import list
Modulo is intended to manage the modules of a single project, but
modules include imports from many sources outside the project. We can
exclude those simply by checking to see if the target of the import is a
module that appears in our import list.

These leaves an edge case to be aware of that we may want to fill at
some point: modules that contain no imports. Technically we could fail
to find a cycle if one of the modules involved in it happen to only use
function from Prelude. This is a very small edge case that, so I've
decided to let it be for now.