Calculates dependency graph for modules trees
GHC already makes sense that modules themselves do not have cyclic
dependencies between them (otherwise it can't compile the code!). What
we're actually interested in is the dependency between the modules when
they are thought of as a tree hierarchy. To the language itself, modules
live in a completely flat namespace -- the hierarchy implied by the dot
notation is for file organization only.

This builds up a view of the dependencies between the levels of the
hierarchy by converting the import states into the implied dependencies
produced by rolling up the imports to the imagined hierarchy of modules
implied by the dot notation. This changes main to now print out this
graph rather than the raw imports.

I've left one edge case to fix later -- modules which actually import
their parents directly. Because we want to let modules import their
sibling which live under the same node, we currently filter out parent
dependencies when generating the graph. (e.g. so A.B.C importing A.B.D
doesn't result in a A.B.C having an implied dependency on A.B). However
if A.B.C where to import A.B directly, we currently also filter that
out. I've noted this in the readme so we can fix it later.