Preserves tree structure in TreeName values
`treeNamesForModule` does work to extract the tree structure from the
names, but that work was not being captured in the `TreeName` values.
Instead, `treeNamesForModule` was essentially forced to throw the work
away (putting the `.`s back in) to build the `String` expected by
`TreeName`.

This changes the `TreeName` type's internal structure to preserve the
knowledge about the tree structure.