Adds encapsulated module tree checking
An "encapsulated module tree" as tree where the top node is intended to
export a public interface for callers to user rather than callers
importing the modules contained within the tree directly.

Currently modulint requires a specific tree structure to enforce this --
namely that the public facing module is the root of the tree.