Add package version constraints, particularly for Dhall.
Compiling aganist an older version of dhall fails, so I added some
package constraints. These are based on some semantic versioning
guesswork, but should serve as a good starting.

I opted _not_ to constraint `bytestring` and `text` because they have
wide ranges and modulint does not use them in as sophisticated way. We
may want to revist that.