Adds support for checking alias uniqueness
This allows to write a rule for allowing or not the aliases used, per
module, to be unique.

Rules allow for three styles:

- No enforcement of alias uniqueness. This is the default, matching
behavior prior to this feature

- Enforcement of all aliases to be unique. With this rule, when
processing a module all of the aliases must be unique, but an
allowlist is given for aliases that are permitted to be duplicated.

For example, the allow list is useful if you want to enforce every
module to have a unique set of alias, except for re-export modules
that may for brevity use the same alias repeatedly.

- Enforcement of certain aliases to be unique. This is a blocklist
style where the given aliases are the only ones checked for
uniqueness.

For example this can be used when you want to disallow clashes for
common aliases. Consider an import of 'Control.Monad' and
'Data.Maybe', they might otherwise be aliased as 'M', but if you want
to enforce them to be distinct then the blocklist allows for checking
only those cases.

A major limitation of this is that the rules are such that they apply
to every module individually and can only be stated as such. The
maximum open import rule provides an example on how we might extend to
support rules per module in the future.This means that the reuse of an
alias, for different import statements, but existing in separate
source modules is not checked. This is a consequence of running as a
plugin where we do not have a good way to build up state. Further,
building up a map for all imports by alias across large projects is
likely to have negative impacts on build speed.