More on work with documentation checks
Example dhall has rule specified to see how documentation rules can
look. An open question is around the use of a generated span for the
"location" of the haddock rule failure.

Further, this work makes a few things clear that need to be addressed.

First is that we clearly have a pattern for rules that have a maximum
amount as well as a mirror of them for minimum amount rules.

Second is that in order to support a diagnostic code in a sane way,
some more thought needs to be given to the way the various rule check
failures are structured.