Update to configuration
Configuration is now more concise and consistent particularly when
applying some rules to any given module and having some for a
specified module.

When processing imports, some attempt is made at looking at each
import for all the various rules configured in a single shot. This
uses DList now to continually append any failures.

Also, this adds some rules for enforcing the @since haddock annotation
on exports.

Finally, this simplifies both the configuration for users and the
implementation by collapsing rules that are different sorts of
"maximum" or "minimum".