Switching to a TOML config instead of Dhall
TOML gives us several advantages:

- Most users are much more likely to be familiar with it.

- Configuration now expressed in a single file; transitions should be
smoother in the face of configuration format changes.

- The 'tomland' package significantly reduces the dependency tree vs
'dhall'. Which, _may_ have a positive impact on runtime speed using
henforcer.

- Documentating the TOML format is easier because there are less
concepts to be concerned with, like functions.

- Errors with 'Dhall' were not great out of the box and 'tomland'
gives us better error messages out of the gate, though we likely still
want to some work here.

TOML does cause us to lose out on some things:

- Network references built in meant it was easily possible to share a
base configuration.

- Dhall is a much more expressive language leading to a less verbose
configuration compared to TOML.

The belief is the tradeoffs for TOML will be a net positive for users.

On 'tomland': The library does not use a typeclass based approach,
which is prefereable and matches somewhat closely to what we had with
Dhall. It also has a relatively small dependency tree and was quick
enough to get the switch completed.