Adding an optional note field to several of the rules, which allows a user to include more context when a rule fails.
Updating the README to better display configuration options.

Also allows for import schemes to be overriden for specified modules.
This will enable users to effectively disallow imports except in
certain modules. A potential use case is a wrapper around some library
that is then not allowed to be imported in the rest of an application.