Rework a bit of the recompilation detection.
Being a bit smarter here should lead to an overall better experience.
This will continue to not parse the entire configuration when no
change has been made. However there is better detection for when the
configuration has been read once and then changed. This should allow
for better consistency on a configuration change.