Add functionality to ignore rules for a specified module
The headline feature allows users to have a rule defined for any
module and then completely ignore that rule, as if it was never
specified, for a specified module. This way rules that may not have
true ways to "undo" the configuration in themselves can be turned off.

This feature remains undocumented for the time being.

Along the way the Config module is broken up, and some common rule
pieces around conditional enforcement are abstracted for consistency.
Although more can be done in this regard. Particularly for alias
uniqueness allowance, encapsulated trees, and tree dependencies.