Adds wildcard matching feature for rules and restructure 'allowedAliasUniqueness'
Now support a 'forPatternModules' top level array of tables with all
of ability to state rules as 'forSpecifiedModules'. Pattern rules will
take precedence over 'forAnyModule' and 'forSpecifiedModules' will
take the highest precedence. Structuring this way allows us to
reasonably add rules that apply to any module, override a subset of
those for a module tree and further override the tree rules for a
specific module.

Also the 'allowedAliasUniqueness' rule is reworked and now has two
acceptable table forms instead of a boolean that changes the behavior
of another key in the configuration. This should give us more clarity
for new users.